+++ datafusion-cli/src/print_format.rs
@@ -23,8 +23,10 @@ use crate::print_options::MaxRows;
 
 use arrow::csv::writer::WriterBuilder;
 use arrow::datatypes::SchemaRef;
+use arrow::error::ArrowError;
 use arrow::json::{ArrayWriter, LineDelimitedWriter};
 use arrow::record_batch::RecordBatch;
+use arrow::util::display::{ArrayFormatter, ValueFormatter};
 use arrow::util::pretty::pretty_format_batches_with_options;
 use datafusion::common::format::DEFAULT_CLI_FORMAT_OPTIONS;
 use datafusion::error::Result;
@@ -208,6 +210,152 @@ impl PrintFormat {
             _ => {}
         }
         Ok(())
+    }
+
+    pub fn process_batch(
+        &self,
+        batch: &RecordBatch,
+        schema: SchemaRef,
+        preview_batches: &mut Vec<RecordBatch>,
+        preview_row_count: &mut usize,
+        preview_limit: usize,
+        precomputed_widths: &mut Option<Vec<usize>>,
+        header_printed: &mut bool,
+        writer: &mut dyn std::io::Write,
+    ) -> Result<()> {
+        if precomputed_widths.is_none() {
+            preview_batches.push(batch.clone());
+            *preview_row_count += batch.num_rows();
+            if *preview_row_count >= preview_limit {
+                let widths =
+                    Self::compute_column_widths(self, preview_batches, schema.clone())?;
+                *precomputed_widths = Some(widths.clone());
+                Self::print_header(self, &schema, &widths, writer)?;
+                *header_printed = true;
+                for preview_batch in preview_batches.drain(..) {
+                    Self::print_batch_with_widths(self, &preview_batch, &widths, writer)?;
+                }
+            }
+        } else {
+            let widths = precomputed_widths.as_ref().unwrap();
+            if !*header_printed {
+                Self::print_header(self, &schema, widths, writer)?;
+                *header_printed = true;
+            }
+            Self::print_batch_with_widths(self, batch, widths, writer)?;
+        }
+        Ok(())
+    }
+
+    pub fn compute_column_widths(
+        &self,
+        batches: &Vec<RecordBatch>,
+        schema: SchemaRef,
+    ) -> Result<Vec<usize>> {
+        let mut widths: Vec<usize> =
+            schema.fields().iter().map(|f| f.name().len()).collect();
+        for batch in batches {
+            let formatters = batch
+                .columns()
+                .iter()
+                .map(|c| ArrayFormatter::try_new(c.as_ref(), &DEFAULT_CLI_FORMAT_OPTIONS))
+                .collect::<Result<Vec<_>, ArrowError>>()?;
+            for row in 0..batch.num_rows() {
+                for (i, formatter) in formatters.iter().enumerate() {
+                    let cell = formatter.value(row);
+                    widths[i] = widths[i].max(cell.to_string().len());
+                }
+            }
+        }
+        Ok(widths)
+    }
+
+    pub fn print_header(
+        &self,
+        schema: &SchemaRef,
+        widths: &Vec<usize>,
+        writer: &mut dyn std::io::Write,
+    ) -> Result<()> {
+        Self::print_border(widths, writer)?;
+
+        let header: Vec<String> = schema
+            .fields()
+            .iter()
+            .enumerate()
+            .map(|(i, field)| Self::pad_cell(field.name(), widths[i]))
+            .collect();
+        writeln!(writer, "| {} |", header.join(" | "))?;
+
+        Self::print_border(widths, writer)?;
+        Ok(())
+    }
+
+    pub fn print_batch_with_widths(
+        &self,
+        batch: &RecordBatch,
+        widths: &Vec<usize>,
+        writer: &mut dyn std::io::Write,
+    ) -> Result<()> {
+        let formatters = batch
+            .columns()
+            .iter()
+            .map(|c| ArrayFormatter::try_new(c.as_ref(), &DEFAULT_CLI_FORMAT_OPTIONS))
+            .collect::<Result<Vec<_>, ArrowError>>()?;
+        for row in 0..batch.num_rows() {
+            let cells: Vec<String> = formatters
+                .iter()
+                .enumerate()
+                .map(|(i, formatter)| Self::pad_value(&formatter.value(row), widths[i]))
+                .collect();
+            writeln!(writer, "| {} |", cells.join(" | "))?;
+        }
+        Ok(())
+    }
+
+    // 辅助函数：打印一行点行，用于表示被省略的行
+    pub fn print_dotted_line(
+        &self,
+        widths: &Vec<usize>,
+        writer: &mut dyn std::io::Write,
+    ) -> Result<()> {
+        // 构造每个单元格，点号左对齐，长度与对应宽度相同
+        let cells: Vec<String> = widths
+            .iter()
+            .map(|&w| format!(" {: <width$} ", ".", width = w))
+            .collect();
+        // 按照 " | " 拼接，并在两侧加上边框符号
+        writeln!(writer, "|{}|", cells.join("|"))?;
+        Ok(())
+    }
+
+    // 辅助函数：打印底部边框
+    pub fn print_bottom_border(
+        &self,
+        widths: &Vec<usize>,
+        writer: &mut dyn std::io::Write,
+    ) -> Result<()> {
+        // 构造每个单元格对应的边框部分，例如 "+------------+"
+        let cells: Vec<String> = widths
+            .iter()
+            .map(|&w| "-".repeat(w + 2)) // 加2可以对齐左右两侧的空格
+            .collect();
+        writeln!(writer, "+{}+", cells.join("+"))?;
+        Ok(())
+    }
+
+    fn print_border(widths: &Vec<usize>, writer: &mut dyn std::io::Write) -> Result<()> {
+        let cells: Vec<String> = widths.iter().map(|&w| "-".repeat(w + 2)).collect();
+        writeln!(writer, "+{}+", cells.join("+"))?;
+        Ok(())
+    }
+
+    fn pad_cell(cell: &str, width: usize) -> String {
+        format!("{:<width$}", cell, width = width)
+    }
+
+    fn pad_value(formatter: &ValueFormatter, width: usize) -> String {
+        let s = formatter.try_to_string().unwrap_or_default();
+        format!("{:<width$}", s, width = width)
     }
 }
