+++ datafusion-cli/src/print_options.rs
@@ -29,6 +29,7 @@ use datafusion::common::DataFusionError;
 use datafusion::error::Result;
 use datafusion::physical_plan::RecordBatchStream;
 
+use datafusion::execution::SendableRecordBatchStream;
 use futures::StreamExt;
 
 #[derive(Debug, Clone, PartialEq, Copy)]
@@ -106,9 +107,123 @@ impl PrintOptions {
         Ok(())
     }
 
+    pub async fn print_table_batch(
+        &self,
+        print_options: &PrintOptions,
+        schema: SchemaRef,
+        stream: &mut SendableRecordBatchStream,
+        max_rows: usize,
+        writer: &mut dyn std::io::Write,
+        now: Instant,
+    ) -> Result<()> {
+        let preview_limit: usize = 1000;
+        let mut preview_batches: Vec<RecordBatch> = vec![];
+        let mut preview_row_count = 0_usize;
+        let mut total_count = 0_usize;
+        let mut precomputed_widths: Option<Vec<usize>> = None;
+        let mut header_printed = false;
+        let mut max_rows_reached = false;
+
+        while let Some(batch) = stream.next().await {
+            let batch = batch?;
+            let batch_rows = batch.num_rows();
+
+            if !max_rows_reached {
+                if total_count < max_rows {
+                    if total_count + batch_rows > max_rows {
+                        let needed = max_rows - total_count;
+                        let batch_to_print = batch.slice(0, needed);
+                        print_options.format.process_batch(
+                            &batch_to_print,
+                            schema.clone(),
+                            &mut preview_batches,
+                            &mut preview_row_count,
+                            preview_limit,
+                            &mut precomputed_widths,
+                            &mut header_printed,
+                            writer,
+                        )?;
+                        if precomputed_widths.is_none() {
+                            let widths = print_options.format.compute_column_widths(
+                                &preview_batches,
+                                schema.clone(),
+                            )?;
+                            precomputed_widths = Some(widths.clone());
+                            if !header_printed {
+                                print_options
+                                    .format
+                                    .print_header(&schema, &widths, writer)?;
+                                header_printed = true;
+                            }
+                            for preview_batch in preview_batches.drain(..) {
+                                print_options.format.print_batch_with_widths(
+                                    &preview_batch,
+                                    &widths,
+                                    writer,
+                                )?;
+                            }
+                        }
+                        if let Some(ref widths) = precomputed_widths {
+                            for _ in 0..3 {
+                                print_options.format.print_dotted_line(widths, writer)?;
+                            }
+                            print_options.format.print_bottom_border(widths, writer)?;
+                        }
+                        max_rows_reached = true;
+                    } else {
+                        print_options.format.process_batch(
+                            &batch,
+                            schema.clone(),
+                            &mut preview_batches,
+                            &mut preview_row_count,
+                            preview_limit,
+                            &mut precomputed_widths,
+                            &mut header_printed,
+                            writer,
+                        )?;
+                    }
+                }
+            }
+            total_count += batch_rows;
+        }
+
+        if !max_rows_reached {
+            if precomputed_widths.is_none() && !preview_batches.is_empty() {
+                let widths = print_options
+                    .format
+                    .compute_column_widths(&preview_batches, schema.clone())?;
+                precomputed_widths = Some(widths);
+                if !header_printed {
+                    print_options.format.print_header(
+                        &schema,
+                        precomputed_widths.as_ref().unwrap(),
+                        writer,
+                    )?;
+                    header_printed = true;
+                }
+            }
+            if let Some(ref widths) = precomputed_widths {
+                print_options.format.print_bottom_border(widths, writer)?;
+            }
+        }
+
+        // 打印执行详情（例如总行数、耗时等）
+        let formatted_exec_details = print_options.get_execution_details_formatted(
+            total_count,
+            print_options.maxrows,
+            now, // 查询开始时间或当前时间
+        );
+        if !print_options.quiet {
+            writeln!(writer, "{}", formatted_exec_details)?;
+        }
+
+        Ok(())
+    }
+
     /// Print the stream to stdout using the specified format
     pub async fn print_stream(
         &self,
+        max_rows: MaxRows,
         mut stream: Pin<Box<dyn RecordBatchStream>>,
         query_start_time: Instant,
     ) -> Result<()> {
@@ -116,6 +231,11 @@ impl PrintOptions {
             return Err(DataFusionError::External(
                 "PrintFormat::Table is not implemented".to_string().into(),
             ));
+        };
+
+        let max_count = match self.maxrows {
+            MaxRows::Unlimited => usize::MAX,
+            MaxRows::Limited(n) => n,
         };
 
         let stdout = std::io::stdout();
@@ -127,21 +247,20 @@ impl PrintOptions {
         while let Some(maybe_batch) = stream.next().await {
             let batch = maybe_batch?;
             row_count += batch.num_rows();
-            self.format.print_batches(
-                &mut writer,
-                batch.schema(),
-                &[batch],
-                MaxRows::Unlimited,
-                with_header,
-            )?;
+            if row_count < max_count || (with_header && row_count > max_count) {
+                self.format.print_batches(
+                    &mut writer,
+                    batch.schema(),
+                    &[batch],
+                    max_rows,
+                    with_header,
+                )?;
+            }
             with_header = false;
         }
 
-        let formatted_exec_details = self.get_execution_details_formatted(
-            row_count,
-            MaxRows::Unlimited,
-            query_start_time,
-        );
+        let formatted_exec_details =
+            self.get_execution_details_formatted(row_count, max_rows, query_start_time);
 
         if !self.quiet {
             writeln!(writer, "{formatted_exec_details}")?;
